scripts/requirements.txt
python-dotenv
supabase
googlemaps
requests
psycopg2-binary
